Refactored Vocab and VocabCollection. Updated Readme for examples
