+++ examples/HealthData/Readme.md
@@ -1,9 +1,9 @@
 ﻿# Health Data Agent
 This example requires GPT-4.
 
-Demonstrates a ***strongly typed*** Chat. Here, you work with an Agent to interactively enter medications or conditions. 
+Demonstrates a ***strongly typed*** Chat: a natural language interface for entering health information. You work with a Health Data Agent to interactively enter your medications or conditions. 
 
-Shows how to use **Agents (Chat) with history** to ***interactively*** acquire information needed for one or more Types from the user: **form filling**. 
+The Health Data Agent shows how strongly typed **Agents with history** could interact with a user to collect information needed for one or more data types ("form filling"). 
 
 ## Target models
 For best and consistent results, use **gpt-4**.
