+++ examples/Math/Readme.md
@@ -1,6 +1,6 @@
 ﻿# Math
 
-The Math example shows how to use **TypeChat.Program** for program generation based on an API schema. This example translates calculations into simple programs given an [`API`](MathAPI.cs) type that can perform the four basic mathematical operations.
+The Math example shows how to use **TypeChat.Program** for program generation based on an API schema. This example translates calculations into simple programs given an [`API`](MathAPI.cs) type that can perform the basic mathematical operations.
 
 ## Target models
 Works with gpt-35-turbo and gpt-4.
@@ -9,7 +9,7 @@ Works with gpt-35-turbo and gpt-4.
 
 Example prompts can be found in [`input.txt`](input.txt).
 
-For example, we could use natural language to describe mathematical operations, and TypeChat.Program will generate a json program for the math API defined in the schema. It will then execute the program using a simple interpreter.
+For example, we could use natural language to describe mathematical operations, and TypeChat.Program will generate a json program for the math API defined in the schema. It will then execute the program.
 
 **Input**:
