+++ src/typechat/Schema/VocabCollection.cs
@@ -5,15 +5,6 @@ using System.Collections;
 namespace Microsoft.TypeChat.Schema;
 
 /// <summary>
-/// A collection of named vocabularies.
-/// Vocabulary collections can be dynamic and load vocabularies at runtime from data bases, files and so on
-/// </summary>
-public interface IVocabCollection : IEnumerable<NamedVocab>
-{
-    NamedVocab? Get(string name);
-}
-
-/// <summary>
 /// An in-memory vocabulary collection
 /// </summary>
 public class VocabCollection : IVocabCollection
